Trying fix

+++ database/migrations/2019_08_04_214011_create_feature_values_table.php
@@ -17,7 +17,7 @@ class CreateFeatureValuesTable extends Migration
             $table->increments('id');
             $table->string('value_string')->nullable();
             $table->integer('value_integer')->nullable();
-            $table->integer('feature_id');
+            $table->bigInteger('feature_id');
             $table->foreign('feature_id')->references('id')->on('features')->onDelete('cascade');
         });
     }
